requirements.txt
flask
flask-cors
requests
python-dotenv
pydantic
feedparser
cryptography
